+++ package.json
@@ -101,6 +101,7 @@
         "sass-loader": "^12.6.0",
         "ts-loader": "^9.3.0",
         "typescript": "^4.6.4",
+        "url-loader": "^4.1.1",
         "webpack-cli": "^4.9.2",
         "webpack-dev-server": "^4.7.4"
     },
