+++ public/index.html
@@ -8,7 +8,8 @@
     </head>
 
     <body>
-        <div id="root"></div>
+        <div id="root">
+        </div>
         <script>
             var exports = {};
         </script>
