+++ public/styles.css
@@ -1,14 +1,57 @@
+@import url('https://fonts.googleapis.com/css?family=Abel|Abril+Fatface|Alegreya|Arima+Madurai|Dancing+Script|Dosis|Merriweather|Oleo+Script|Overlock|PT+Serif|Pacifico|Playball|Playfair+Display|Share|Unica+One|Vibur');
+
 body {
     margin: 0 auto;
     background-image: linear-gradient(
+        180deg,
+        rgba(255, 195, 223, 0.4738270308123249) 0%,
+        rgba(238, 174, 202, 1) 66%,
+        rgba(201, 249, 188, 0.534) 100%
+    );
+
+    /* background-image: linear-gradient(
         -45deg,
         #35c3f3 0%,
         #8b9fe8 20%,
         #e681d8 39%,
         #ffa9a4 76%,
         #fed2ce 100%
-    );
+    ); */
     width: 100%;
+    background-size: 100% 100%;
+    height: 100vh;
+    background-repeat: no-repeat;
+}
+
+.container {
+    width: 100%;
+    height: 100%;
+}
+
+section {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100vw;
+    height: 100vh;
+    color: white;
+}
+
+section#pageOne {
+    background-image: url('https://source.unsplash.com/1lfI7wkGWZ4/1600x900');
+    background-size: 100% 100%;
+}
+
+section#pageTwo {
+    /* background-image: linear-gradient(180deg, rgb(228, 255, 146) 0%, #ffe684 40%, #94afe9 100%); */
+    background-image: url('https://source.unsplash.com/5QgIuuBxKwM/1600x900');
+    background-size: 100% 100%;
+}
+
+section#pageThree {
+    /* background-image: linear-gradient(180deg, #94afe9 0%, #e9a6df 66%, #b57cff 100%); */
+    background-image: url('https://source.unsplash.com/BStW5kYXw4E/1600x900');
+    background-size: 100% 100%;
 }
 
 body a {
@@ -16,6 +59,13 @@ body a {
     text-decoration: none;
     margin: 0 5px;
 }
+
+img {
+    width: 25%;
+    margin-top: -75px;
+    margin-right: 300px;
+}
+
 #main {
     color: #353535;
     margin: 0 auto;
@@ -26,38 +76,128 @@ body a {
     justify-content: center;
 }
 #navbar {
-    color: #353535;
-    text-align: right;
-    background-color: rgba(255, 255, 255, 0.2);
+    color: transparent;
+    background-color: transparent;
     width: 100%;
-    height: 48px;
+    height: 100px;
+    display: flex;
+    flex-direction: row;
+    align-items: flex-start;
+    margin-top: 30px;
+    justify-content: space-between;
+    position: fixed;
 }
+.logoText {
+    font-size: 150%;
+    font-family: 'Playfair Display', serif;
+    color: #ffffff;
+    text-align: center;
+    margin-left: 50px;
+    margin-top: -20px;
+}
+
+.text {
+    font-size: 250%;
+    font-family: 'Playfair Display', serif;
+    color: #3e403f;
+}
+
 .linkBtn {
+    appearance: none;
+    background-color: transparent;
+    border-width: 0;
+    box-sizing: border-box;
+    color: #ffffff;
+    cursor: pointer;
     display: inline-block;
-    padding: 15px 30px;
-    transition: 0.5s;
+    font-family: 'Playfair Display', serif;
+    font-size: 16px;
+    letter-spacing: 0;
+    line-height: 1em;
+    margin: 0;
+    opacity: 1;
+    outline: 0;
+    padding: 1.5em 2.2em;
+    position: relative;
+    text-align: center;
+    text-decoration: none;
+    text-rendering: geometricprecision;
+    text-transform: uppercase;
+    transition: opacity 300ms cubic-bezier(0.694, 0, 0.335, 1),
+        background-color 100ms cubic-bezier(0.694, 0, 0.335, 1),
+        color 100ms cubic-bezier(0.694, 0, 0.335, 1);
+    user-select: none;
+    -webkit-user-select: none;
+    touch-action: manipulation;
+    vertical-align: baseline;
+    white-space: nowrap;
+    margin-right: 50px;
+    font-weight: 500;
 }
-.linkBtn:hover {
-    background-color: #a496e3;
+
+.linkBtn:before {
+    animation: opacityFallbackOut 0.5s step-end forwards;
+    backface-visibility: hidden;
+    background-color: #ffffff78;
+    clip-path: polygon(-1% 0, 0 0, -25% 100%, -1% 100%);
+    content: '';
+    height: 100%;
+    left: 0;
+    position: absolute;
+    top: 0;
+    transform: translateZ(0);
+    transition: clip-path 0.5s cubic-bezier(0.165, 0.84, 0.44, 1),
+        -webkit-clip-path 0.5s cubic-bezier(0.165, 0.84, 0.44, 1);
+    width: 100%;
 }
+
+.linkBtn:hover:before {
+    animation: opacityFallbackIn 0s step-start forwards;
+    clip-path: polygon(0 0, 101% 0, 101% 101%, 0 101%);
+}
+
+.linkBtn span {
+    z-index: 1;
+    position: relative;
+}
+
 #footer {
-    color: #353535;
+    color: #1d1946;
     text-align: center;
     padding: 20px 0;
-    height: 60px;
-    background-color: rgba(255, 255, 255, 0.2);
+    height: 50px;
+    /* background-color: rgba(255, 255, 255, 0.2); */
+    position: fixed;
+    bottom: 0px;
+    width: 100%;
 }
-.form-wrapper input {
+.form-wrapper {
     /* margin: 0 auto; */
     display: block;
 }
-.input {
+.email,
+.password {
+    width: 240px;
+    height: 30px;
+
+    margin-top: 2%;
+    padding: 15px;
+
+    font-size: 16px;
+    font-family: 'Abel', sans-serif;
+    color: #5e6472;
+
+    outline: none;
     border: none;
+
+    border-radius: 0px 5px 5px 0px;
+    transition: 0.2s linear;
+}
+
+.email:focus,
+.password:focus {
+    transform: translateX(-2px);
     border-radius: 5px;
-    text-align: center;
-    width: 200px;
-    padding: 5px;
-    margin: 20px auto;
 }
 .btn {
     color: #353535;
@@ -70,15 +210,24 @@ body a {
     transition: 0.5s;
 }
 .btn:hover {
-    background-color: #a496e3;
+    background-color: #e8a4ffc1;
 }
 .box {
-    margin: 130px 0;
-    color: #353535;
-    padding: 30px;
-    border-radius: 15px;
-    background-color: rgba(255, 255, 255, 0.3);
+    margin: 2% auto 10% auto;
+    text-align: center;
+    width: 450px;
+    min-height: 500px;
+    height: auto;
+    border-radius: 5px;
+    box-shadow: 0 9px 50px hsla(20, 67%, 75%, 0.31);
+    padding: 2%;
+    background-image: linear-gradient(
+        -225deg,
+        rgb(255, 202, 225) 50%,
+        rgba(201, 249, 188, 0.534) 50%
+    );
 }
+
 .box .transferBtn {
     color: #0094c6;
     font-size: 15px;
@@ -86,13 +235,37 @@ body a {
     margin: 0;
     background-color: rgba(255, 255, 255, 0);
     transition: 0.5s;
+    text-decoration: underline;
 }
 .transferBtn:hover {
     color: #35c3f3;
     background-color: rgba(255, 255, 255, 0);
 }
+
+.formBtn {
+    display: inline-block;
+    color: #252537;
+
+    width: 280px;
+    height: 50px;
+
+    padding: 0 20px;
+    background: #fff;
+    border-radius: 5px;
+
+    outline: none;
+    border: none;
+
+    cursor: pointer;
+    text-align: center;
+    transition: all 0.2s linear;
+
+    margin: 7% auto;
+    letter-spacing: 0.05em;
+}
+
 .paragraph {
-    margin: 0;
+    margin: 2rem auto 1rem;
     display: inline-block;
 }
 .app h1 {
